requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.29
httpx>=0.27
pyjwt>=2.8
cryptography>=42
python-dotenv>=1.0
